refactor(packages): remove @slate-serializers/tests dep
