+++ packages/react/tsconfig.spec.json
@@ -16,5 +16,5 @@
     "src/**/*.test.jsx",
     "src/**/*.spec.jsx",
     "src/**/*.d.ts"
-  ]
+, "../tests/src/lib/react/components.spec.tsx"  ]
 }
