+++ packages/utilities/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/style-object';
 export * from './lib/utilities';
-export * from '../../html/src/lib/utilities/blocks';
